Updated to API version 1.7.14073.0

+++ MozuApi/Contracts/Content/MOZUPropertyTypeCollection.h
@@ -26,7 +26,7 @@
 /**
 The number of pages returned based on the startIndex and pageSize values specified. This value is system-supplied and read-only.
 */
-@property(nonatomic) long pageCount;
+@property(nonatomic) NSInteger pageCount;
 
 /**
 The number of results to display on each page when creating paged results from a query. The maximum value is 200.
@@ -41,7 +41,7 @@ When creating paged results from a query, this value indicates the zero-based of
 /**
 The number of results listed in the query collection, represented by a signed 64-bit (8-byte) integer. This value is system-supplied and read-only.
 */
-@property(nonatomic) long totalCount;
+@property(nonatomic) NSInteger totalCount;
 
 /**
 An array list of objects in the returned collection.
